+++ frontend/app/App.js
@@ -1,126 +1,14 @@
-import React, { useState } from "react";
-import { View, Image, Platform } from "react-native";
+import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { Canvas } from "react-three-fiber";
-import { ThemeProvider, Header } from "react-native-elements";
+import { ThemeProvider } from "react-native-elements";
 
-import Model from "./src/Model";
-import Menu from "./src/Menu";
-import { SteelFrame, LoadingIndicator } from "./src/models";
-
-function BasementScreen() {
-  return (
-    <Canvas camera={{ position: [0, 0, 50] }}>
-      <ambientLight intensity={0.5} />
-      <spotLight intensity={0.8} position={[300, 300, 400]} />
-      <LoadingIndicator />
-    </Canvas>
-  );
-}
-
-function StrongFloorScreen() {
-  return (
-    <Canvas camera={{ position: [0, 0, 50] }}>
-      <ambientLight intensity={0.5} />
-      <spotLight intensity={0.8} position={[300, 300, 400]} />
-      <LoadingIndicator />
-    </Canvas>
-  );
-}
-
-async function fetchData(
-  sensorPackage,
-  dataType,
-  averagingWindow,
-  startTime,
-  endTime
-) {
-  try {
-    const response = await fetch(
-      `http://129.169.72.175/fbg/${sensorPackage}/${dataType}/?averaging-window=${averagingWindow}&start-time=${startTime}&end-time=${endTime}`,
-      {
-        method: "GET",
-        headers: { "media-type": "application/json" },
-      }
-    );
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    console.error(error);
-  }
-}
-
-function SteelFrameScreen() {
-  const [data, setData] = useState([]);
-  const [mode, setMode] = useState(0);
-  const [modelModeEnabled, setModelModeEnabled] = useState(true);
-  const [dataType, setDataType] = useState("str");
-  const [isLoading, setIsLoading] = useState(false);
-
-  async function refresh(dataType, averagingWindow, startTime, endTime) {
-    setIsLoading(true);
-    // Try fetching data
-    try {
-      setData(
-        await fetchData(
-          "steel-frame",
-          dataType,
-          averagingWindow,
-          startTime.toISOString(),
-          endTime.toISOString()
-        )
-      );
-      // Set the type of the fetched data
-      setDataType(dataType);
-      // Enable/disable the model mode button
-      if (dataType == "raw") {
-        setMode(1);
-        setModelModeEnabled(false);
-      } else {
-        setModelModeEnabled(true);
-      }
-    } catch (error) {
-      console.error(error);
-    }
-    setIsLoading(false);
-  }
-
-  function renderVisualisation() {
-    if (mode == 0) {
-      return (
-        <Model
-          file={require("./assets/models/steel-frame.glb")}
-          data={data}
-          dataType={dataType}
-        >
-          {({ localUri, rotation, sensorColours }) => (
-            <SteelFrame
-              localUri={localUri}
-              rotation={rotation}
-              sensorColours={sensorColours}
-            />
-          )}
-        </Model>
-      );
-    } else if (mode == 1) {
-      return <View style={{ flex: 5 }} />;
-    } else return null;
-  }
-
-  return (
-    <View style={{ flex: 1, flexDirection: "row" }}>
-      {renderVisualisation()}
-      <Menu
-        mode={mode}
-        setMode={setMode}
-        modelModeEnabled={modelModeEnabled}
-        isLoading={isLoading}
-        refresh={refresh}
-      />
-    </View>
-  );
-}
+import {
+  BasementScreen,
+  StrongFloorScreen,
+  SteelFrameScreen
+} from "./src/screens";
+import { Header } from "./src/components";
 
 const Tab = createBottomTabNavigator();
 
@@ -128,8 +16,8 @@ const theme = {
   Text: {
     style: {
       fontSize: 20,
-      fontFamily: Platform.select({ default: "Roboto", ios: "Helvetica" }),
-    },
+      fontFamily: Platform.select({ default: "Roboto", ios: "Helvetica" })
+    }
   },
   Divider: { style: { margin: 4 } },
   Button: {
@@ -137,50 +25,20 @@ const theme = {
       paddingTop: 5,
       paddingBottom: 5,
       paddingLeft: 10,
-      paddingRight: 10,
+      paddingRight: 10
     },
     buttonStyle: { borderWidth: 1 },
     titleStyle: {
-      fontSize: 16,
-    },
+      fontSize: 16
+    }
   },
-  ButtonGroup: { textStyle: { fontSize: 16 } },
+  ButtonGroup: { textStyle: { fontSize: 16 } }
 };
 
 export default function App() {
   return (
     <ThemeProvider theme={theme}>
-      <Header
-        containerStyle={{
-          height: 70,
-          paddingVertical: 20,
-          paddingHorizontal: 20,
-          borderBottomColor: "#404040",
-        }}
-        statusBarProps={{
-          hidden: true,
-        }}
-        placement="left"
-        backgroundColor="#404040"
-        leftComponent={
-          <Image
-            source={require("./assets/images/logo.png")}
-            style={{ width: 75, height: 55 }}
-          />
-        }
-        centerComponent={
-          <Image
-            source={require("./assets/images/title.png")}
-            style={{ width: 60, height: 10 }}
-          />
-        }
-        rightComponent={
-          <Image
-            source={require("./assets/images/cambridge.png")}
-            style={{ width: 100, height: 20 }}
-          />
-        }
-      />
+      <Header />
       <NavigationContainer>
         <Tab.Navigator>
           <Tab.Screen name="Basement" component={BasementScreen} />
