+++ frontend/app/src/components/index.js
@@ -1,0 +1,4 @@
+export { default as Header } from "./Header";
+export { default as Menu } from "./Menu";
+export { default as Model } from "./Model";
+export { default as Modal } from "./Modal";
