+++ frontend/app/src/screens/index.js
@@ -1,0 +1,3 @@
+export { default as BasementScreen } from "./BasementScreen";
+export { default as StrongFloorScreen } from "./StrongFloorScreen";
+export { default as SteelFrameScreen } from "./SteelFrameScreen";
