+++ frontend/app/src/utils/index.js
@@ -1,0 +1,1 @@
+export { default as fetchData } from "./fetchData";
